deployment_pipeline/requirements.txt
evidently==0.4.15
google-cloud-storage==2.14.0
requests==2.31.0
python-dotenv==1.0.1